implement exam-management for the admins and implemented smaller changes and fixes
